Hilbert {pbdDMAT}	R Documentation

Generate Hilbert Matrices

Description

Methods for constructing Hilbert matrices: H[i,j] = 1/(i+j-1)

Usage

Hilbert(n, type = "matrix", ..., bldim = .pbd_env$BLDIM,
  ICTXT = .pbd_env$ICTXT)

Arguments

n
number of rows and columns.

type
"matrix" or "ddmatrix".

...
Additional arguments.

bldim
blocking dimension.

ICTXT
BLACS context number.

Details

This constructs the square Hilbert matrix of order n. The return is either a matrix or a distributed matrix depending on the argument type=.

Value

Returns a matrix or a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

dx <- Hilbert(100, type="ddmatrix")

print(dx)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]