condnums {pbdDMAT}	R Documentation

Compute or estimate the Condition Number of a Distributed Matrix

Description

Computes or estimates the condition number.

Usage

## S3 method for class 'ddmatrix'
kappa(z, exact = FALSE, norm = NULL, method = c("qr",
  "direct"), ...)

## S4 method for signature 'ddmatrix'
rcond(x, norm = c("O", "I", "1"), triangular = FALSE,
  ...)

Arguments

exact
logical. Determines whether exact condition number or approximation should be computed.

norm
character. Determines which matrix norm is to be used.

method
character. Determines the method use in computing condition number.

...
Extra arguments.

x, z
numeric distributed matrices.

triangular
logical. If true, only the lower triangle is used.

Value

Returns a number.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

comm.set.seed(diff=T)
x <- ddmatrix("rnorm", 10, 10, bldim=2)

cn <- rcond(x)

comm.print(cn)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]