pbdDMAT Control {pbdDMAT}	R Documentation

Some default parameters for pbdDMAT.

Description

This set of controls is used to provide default values in this package.

Format

Objects contain several parameters for communicators and methods.

Details

The default blocking dmat_opts$BLDIM is c(16,16), which results in a 16 by 16 blocking dimension for distributed matrices. Any time a function takes the bldim= argument, it will default to this value unless the user specifies an alternative.

The default ICTXT is 0. This is the full 2-dimensional processor grid.

[Package pbdDMAT version 0.4-0 Index]