ddmatrix-prcomp {pbdDMAT}	R Documentation

Principal Components Analysis

Description

Performs the principal components analysis.

Usage

prcomp(x, ...)

## S4 method for signature 'ddmatrix'
prcomp(x, retx = TRUE, center = TRUE, scale. = FALSE,
  tol = NULL, ...)

Arguments

x
numeric distributed matrix.

...
Ignored.

retx
logical, indicates whether the rotated variables should be returned

center
logical value, determines whether or not columns are zero centered

scale.
logical value, determines whether or not columns are rescaled to unit variance

tol
a value indicating the magnitude below which components should be omitted. (Components are omitted if their standard deviations are less than or equal to tol times the standard deviation of the first component.) With the default null setting, no components are omitted. Other settings for tol could be tol = 0 or tol = sqrt(.Machine$double.eps), which would omit essentially constant components

Details

prcomp() performs the principal components analysis on the data matrix by taking the SVD. Sometimes core R and pbdDMAT will disagree slightly in what the rotated variables are because of how the SVD is caluclated.

Value

Returns a list.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

comm.set.seed(diff=T)

x <- ddmatrix("rnorm", 10, 10)

y <- prcomp(x)
comm.print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]