ddmatrix-solve {pbdDMAT}	R Documentation

Solve

Description

Solving linear systems and matrix inversion for distributed matrices with R-like syntax, with calculations performed by the PBLAS and ScaLAPACK libraries.

Usage

## S4 method for signature 'ddmatrix,ANY'
solve(a)

## S4 method for signature 'ddmatrix,ddmatrix'
solve(a, b)

Arguments

a, b
numeric distributed matrices. Here, a and b must be on the same BLACS context and have the same blocking dimension.

Details

Extensions of R linear algebra functions.

Value

solve() solves systems and performs matrix inversion when argument b= is missing.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

x <- ddmatrix(1:9, 3)

y <- solve(t(A) %*% A)
print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]