extract {pbdDMAT}	R Documentation

Extract or Replace Parts of a Distributed Matrix

Description

Operators to extract or replace parts of a distributed matrix.

Usage

## S4 method for signature 'ddmatrix'
x[i, j, ICTXT]

Arguments

x
numeric distributed matrix.

i, j
indices specifying elements to extract or replace. Indices can be numeric, character, empty, or NULL. number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

ICTXT
optional BLACS context number for output

...
additional arguments.

Details

[ can be used to extract/replace for a distributed matrix exactly as you would with an ordinary matrix.

The functions rely on reblocking across different BLACS contexts. If i is not empty, then the input distributed matrix will be redistributed along context 1, where extracting/deleting rows does not destroy block-cyclicality. Likewise, if j is not empty, then the input distributed matrix will be redistributed along context 2. When extraction is complete, the matrix will be redistributed across its input context.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

x <- ddmatrix(1:9, 3, bldim=2)

y <- x[, -1]
y <- head(y, 2)
y

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]