getLocal {pbdDMAT}	R Documentation

getLocal

Description

getLocal

Usage

getLocal(x, gi, gj, all.rank = TRUE, gridinfo)

Arguments

x
A distributed matrix.

gi, gj
Global row and column indices, respectively.

all.rank
Logical; if TRUE, then all processes will hold the desired value on exit. Otherwise, only the process who owns the local value returns this value, while every other process returns NULL.

gridinfo
An optional parameter; each local data lookup requires the data contained in gridinfo(ICTXT(x)). So you may specify it yourself (and if you are making many function calls, this is preferable performance-wise), or the lookup will be performed for you.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT)
init.grid()

x <- ddmatrix(1:100, 10, bldim=c(2, 2))

val <- getLocal(x, 5, 1)
comm.print(val, all.rank=TRUE)

val <- getLocal(x, 5, 1, all.rank=FALSE)
comm.print(val, all.rank=TRUE)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]