headsortails {pbdDMAT}	R Documentation

Head and Tail of a Distributed Matrix

Description

The functions rely on reblocking across different BLACS contexts. If i is not empty, then the input distributed matrix will be redistributed along context 1, where extracting/deleting rows does not destroy block-cyclicality. Likewise, if j is not empty, then the input distributed matrix will be redistributed along context 2. When extraction is complete, the matrix will be redistributed across its input context.

Usage

## S3 method for class 'ddmatrix'
head(x, n = 6L, ...)

## S3 method for class 'ddmatrix'
tail(x, n = 6L, ...)

Arguments

x
numeric distributed matrix.

n
a single integer. If positive, size for the resulting object: number of elements for a vector (including lists), rows for a matrix or data frame or lines for a function. If negative, all but the n last/first number of elements of x.

...
additional arguments.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

x <- ddmatrix(1:9, 3, bldim=2)

y <- head(y, 2)
print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]