insert {pbdDMAT}	R Documentation

Directly Insert Into Distributed Matrix Submatrix Slot

Description

Allows you to directly replace the submatrix of a distributed matrix.

Usage

## S4 replacement method for signature 'ddmatrix,ANY,ANY,ANY'
x[i, j, ...] <- value

## S4 replacement method for signature 'ddmatrix,ANY,ANY,ddmatrix'
x[i, j, ...] <- value

Arguments

x
numeric distributed matrix.

i, j
global integer indices.

...
Additional arguments.

value
replacement value. Can be a global vector or a ddmatrix.

Details

[<- allows the user to insert values into a distributed matrix in exactly the same way one would with an ordinary matrix. The indices here are global, meaning that x[i, j] refers to the (i, j)'th element of the "full", global matrix, and not necessarily the (i, j)'th element of the local submatrix.

On the other hand, submatrix<- is different. It is basically syntactic sugar for:

x@Data <- newMatrix

It does not alter the distributed matrix x's dim or bldim. It does adjust the ldim automatically. However, using this can be dangerous. It is merely provided to give consistent behavior with the submatrix() function.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

x <- ddmatrix(1:9, 3, bldim=2)

x[1, ] <- 0
comm.print(submatrix(x), all.rank=T)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]