math {pbdDMAT}	R Documentation

Miscellaneous Mathematical Functions

Description

Binary operations for distributed matrix/distributed matrix and distributed matrix/vector operations.

Usage

## S4 method for signature 'ddmatrix'
sqrt(x)

## S4 method for signature 'ddmatrix'
abs(x)

## S4 method for signature 'ddmatrix'
exp(x)

## S4 method for signature 'ddmatrix'
log(x, base = exp(1))

## S4 method for signature 'ddmatrix'
log2(x)

## S4 method for signature 'ddmatrix'
log10(x)

## S4 method for signature 'ddmatrix'
log1p(x)

## S4 method for signature 'ddmatrix'
sin(x)

## S4 method for signature 'ddmatrix'
cos(x)

## S4 method for signature 'ddmatrix'
tan(x)

## S4 method for signature 'ddmatrix'
asin(x)

## S4 method for signature 'ddmatrix'
acos(x)

## S4 method for signature 'ddmatrix'
atan(x)

## S4 method for signature 'ddmatrix'
sinh(x)

## S4 method for signature 'ddmatrix'
cosh(x)

## S4 method for signature 'ddmatrix'
tanh(x)

Arguments

x
numeric distributed matrix

base
a positive number: the base with respect to which logarithms are computed. Defaults to e='exp(1)'.

Details

Performs the miscellaneous mathematical calculation on a distributed matrix.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- sqrt(abs(log(x/10)))
comm.print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]