matmult {pbdDMAT}	R Documentation

Matrix Multiplication

Description

Multiplies two distributed matrices, if they are conformable.

Usage

## S4 method for signature 'ddmatrix,ddmatrix'
x %*% y

## S4 method for signature 'ddmatrix'
crossprod(x, y = NULL)

## S4 method for signature 'ddmatrix'
tcrossprod(x, y = NULL)

Arguments

x, y
numeric distributed matrices

Details

x and y must be conformable, on the same BLACS context, but they need not be blocked with the same blocking dimension. The return will default to the blocking dimension of x.

If you need to use x and y with differing blocking dimensions and you want the return to have blocking different from that of x, then use the function base.rpdgemm().

The crossprod() and tcrossprod() functions behave exactly as their R counterparts.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- x %*% x
print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]