pbdDMAT-package {pbdDMAT}	R Documentation

Distributed Matrix Methods

Description

A package for dense distributed matrix computations. Includes the use of PBLAS and ScaLAPACK libraries via pbdSLAP, communicating over MPI via the BLACS library and pbdMPI.

Details

Package: 	 pbdDMAT

Type: 	 Package

License: 	 GPL

LazyLoad: 	 yes

This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).

Author(s)

Drew Schmidt schmidt AT math.utk.edu, Wei-Chen Chen, George Ostrouchov, and Pragneshkumar Patel, with contributions from R Core team (some wrappers taken from the base and stats packages).

References

Programming with Big Data in R Website: http://r-pbd.org/

[Package pbdDMAT version 0.4-0 Index]