rounding {pbdDMAT}	R Documentation

Rounding of Numbers

Description

Extensions of R rounding functions for distributed matrices.

Usage

## S4 method for signature 'ddmatrix'
round(x, digits = 0)

## S4 method for signature 'ddmatrix'
ceiling(x)

## S4 method for signature 'ddmatrix'
floor(x)

Arguments

x
numeric distributed matrix

digits
integer indicating the number of decimal places (round()) or significant digits (signif()) to be used. Negative values are allowed (see 'Details').

Details

Rounding to a negative number of digits means rounding to a power of ten, so for example round(x, digits = -2) rounds to the nearest hundred.

Value

Returns a distributed matrix.

Examples

## Not run: 
# Save code in a file "demo.r" and run with 2 processors by
# > mpiexec -np 2 Rscript demo.r

library(pbdDMAT, quiet = TRUE)
init.grid()

# don't do this in production code
x <- matrix(1:9, 3)
x <- as.ddmatrix(x)

y <- ceiling(x/3)
print(y)

finalize()

## End(Not run)

[Package pbdDMAT version 0.4-0 Index]