sparsity {pbdDMAT}	R Documentation

Sparsity of Matrix Objects

Description

Determine the sparsity of a matrix, distributed, dense, or otherwise.

Usage

## S4 method for signature 'matrix'
sparsity(x, count = "zero", out = "count",
  tol = .Machine$double.eps)

## S4 method for signature 'vector'
sparsity(x, count = "zero", out = "count",
  tol = .Machine$double.eps)

## S4 method for signature 'dmat'
sparsity(x, count = "zero", out = "count",
  tol = .Machine$double.eps)

Arguments

x
numeric matrix

count
character; options are "zero" and "other". The former counts the number of zeros, while the latter counts the number of non-zeros ('other' elements).

out
character; options are "count", "proportion", and "percent". This determines whether a pure count, proportion of count elements in the matrix, or percentage of count elements in the matrix.

tol
numeric; the tolerance for numerical zero. This is ignored if the input data is integer/logical.

Details

The sparsity count of a matrix is returned.

[Package pbdDMAT version 0.4-0 Index]